show that status column now
